float porting

+++ MachineLearning.Data/Entry/BinaryDataEntry.cs
@@ -2,7 +2,7 @@ using System.Text;
 
 namespace MachineLearning.Data.Entry;
 
-public sealed record BinaryDataEntry(double[] Input, bool Expected) : DataEntry<double[], bool>(Input, Expected)
+public sealed record BinaryDataEntry(float[] Input, bool Expected) : DataEntry<float[], bool>(Input, Expected)
 {
     public override string ToString()
     {
